Optimize ProgressEstimator
File and Dir stats are incremental counters, so they aren't optimizable without introducing a bunch of clutter, so those are left as-is.

The ByteStats though can have a wide range of values being added to it. So if nothing is listening for the updates, there is no need to perform the addition calculation.

Introduced a Queuing mechanism to defer the addition of the Byte calculation until its first requested. From that point on, the addition will occur so that INotifyPropertyChange can be triggered.