Update RoboQueue.cs
I encountered a hang where it would indefinitely wait on line 651, while the underlying Command was awaiting activation.
I believe this was due in part to WPF binding events.
The fix as far here is to await the task until it has been activated, then raise all the events declaring it as such.