Re-Write ObservableList
This file was originally derived from List<T>, but that posed quite a few problems for consumers.
Namely, all interfaces were interacting with the underlying list, and notifications would not be sent via the INotifyCollectionChanged event.

This re-write removes the object being derived from list, and instead add explicit casting to access the underlying list without notifications.
All of the interfaces and methods used by the list have been implemented in the re-write to retain any functionality for consumers.