Updated ObservableListTester
Moved ObservableListTester out of the backupapp and into the unit testing project.
Commented out ObservableListTester since it requires a bit of rework and its already been tested manually.

Essentially, the old method worked for AddRange() in the tester, but once it was put into an application, observers really didn't like the mass notifications, hence previously rewriting it for the foreach loop. This also goes for some of the other methods that modify the list.
As such, the tests need to be re-written for the new logic.