Changes to make custom implementation of IRoboCommand easier
My project has an object called a 'SyncedFile'. I have many of these objects, all in various folders that all get brought into a single destination.
As such, I created a custom IRoboCommand to facilitate this. But needed the estimators and relevant args made public to do so.

After all, no sense in having an interface that can't be implemented.