Fix for 163 (#176)
Fixes #163 - System.NotSupportedException: 'No data is available for encoding 437. For information on defining a custom encoding, see the documentation for the Encoding.RegisterProvider method.'

Increment version number as recent commits also include #170 support for .NET 6.0 authentication, using WindowsIdentity.RunImpersonated API, which is not supported on netstandard2.1 - and - Added PlatformNotSupportedException when trying to use authentication on unsupported target frameworks.